splicing-pattern-analyzer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eukaryotic-splicing-analyzer"
version = "1.0.0"
description = "Splicing dynamics analyzer for eukaryotic genes (TP53 model)"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "biopython>=1.83",
    "rich>=13.7",
    "python-dotenv>=1.0",
]

[project.scripts]
eukaryotic-splicing-analyzer = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]